Fix coding errors that prevent display of GlomModel data in the TreeView
The GlomModel for the MainWindow was allocated on the stack, and not the
heap, in the MainWindow constructor. Also, the table names were being
interpreted as ASCII and not UTF-8. A new private member was added to the
GlomModel class for the header data, and an incorrect conditional check
was fixed. Some comments were added.