Show a table dialog when a table name is double-clicked
Change the doubleClicked handler of the QTreeView in the main window to
show a dialog cantaining a QTableView of the double-clicked table. The
table model is currently blank, as the database connetion is currently
unused. The dialog is non-modal, so a boolean flag to determine the
current state of the dialog was added (table_model_opened).