Made ErrorReporter dependency in GlomLayoutModel dynamic
* src/glom_document.[h|cc]: Forward the MainWindow’s error reporter to
the GlomLayoutModel that is to be created, or report failure.

* src/glom_layout_model.[h|cc]: Extended ctor with an ErrorReporter
argument so that the model can report errors it cannot handle properly
itself.

* src/main_window.[h|cc]: Added an ErrorReporter member to MainWindow
which is connected to the receiveError slot. It allows every component
using this ErrorReporter instance to propagate errors back to the
MainWindow, where an error dialog will be shown and the error is
handled.