Revert "Revert "main(): Check for a QtSQL database driver.""
The comment in the code explains why the check is in main().
This is based on the premise that all versions of Qlom should
support PostgreSQL-based .glom files (the default), so a user
would experience hard-to-understand problems if allowed to get
past this point. Glom has the same check (for libgda providers).

This reverts commit c4e051894b34c459c6e6d11824d78234cddbcfc0.