Added sortable columns
* src/list_layout_model.cc: Sorting by a column means re-quering the database,
completely bypassing the extra setup of our model. Since we added an extra
column *after* executing the query, the sorting mechanism is not aware of
that extra column, hence it would get lost on sorting. The work-around here
is to query a place-holder field when building the query which can then be
used for the actions column by the view. The basic idea was that the model's
column count is not allowed to change once plugged into a view.

* src/layout_delegates.[h|cc]: Silence the text output for button delegates,
regardless what is contained in the model at the given index.

* src/gui/main_window.[h|cc]: Implement sortable columns.